requirements.txt
fastapi
uvicorn
pydantic
onnxruntime
pillow
numpy
